add sqlite

+++ migrations/dlu/3_add_performance_cost.sql
@@ -1,1 +1,1 @@
-ALTER TABLE properties ADD COLUMN performance_cost DOUBLE(20, 15) DEFAULT 0.0;
+ALTER TABLE properties ADD COLUMN performance_cost DOUBLE DEFAULT 0.0;
